// src/enums/error.ts
export enum StatusCode {
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 204,
  MOVED_PERMANENTLY = 301,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  REQUEST_TIMEOUT = 408,
  GONE = 410,
  PAYLOAD_TOO_LARGE = 413,
  UNSUPPORTED_MEDIA_TYPE = 415,
  INTERNAL_SERVER_ERROR = 500,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503
}

export enum ERROR_CODE {
  NOT_FOUND = 'NOT_FOUND',
  INVALID_REQUEST = 'INVALID_REQUEST',
  INVALID_MANDATORY_FIELDS = 'INVALID_MANDATORY_FIELDS',
  UNEXPECTED_ERROR = 'UNEXPECTED_ERROR', // do not use this when create AppError
  UNAUTHORIZED = 'UNAUTHORIZED',
  INCORRECT_FIELD = 'INCORRECT_FIELD',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  EMAIL_EXISTS = 'EMAIL_EXISTS',
  PHONE_NUMBER_EXISTS = 'PHONE_NUMBER_EXISTS',
  PHONE_VALIDATION_EXISTS = 'PHONE_VALIDATION_EXISTS',
  SIGNATURE_VALIDATION_EXISTS = 'SIGNATURE_VALIDATION_EXISTS',
  EMAIL_WAS_VERIFIED = 'EMAIL_WAS_VERIFIED',
  VERIFY_CODE_WAS_EXPIRED = 'VERIFY_CODE_WAS_EXPIRED',
  PHONE_NUMBER_OR_EMAIL_IN_BLACKLISTED = 'PHONE_NUMBER_OR_EMAIL_IN_BLACKLISTED'
}